Fix mainHolder and footer
